requirement.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-anymail==14.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
